lint

+++ server/routers/usersRouter.js
@@ -138,29 +138,22 @@ router.patch(
     },
 );
 
-
-router.delete("/api/users/:id", isLoggedIn, isAccessingOwnUser, (req, res) => {
-
+router.delete('/api/users/:id', isLoggedIn, isAccessingOwnUser, (req, res) => {
     const userId = Number(req.params.id);
 
-    try{
+    try {
+        db.prepare('DELETE FROM users WHERE id = ?').run(userId);
 
-        db.prepare("DELETE FROM users WHERE id = ?")
-        .run(userId);
+        req.session.destroy();
 
-        req.session.destroy()
-
-        return res.status(200).send({ data: { successMessage: "User deleted"}})
-
-    }catch(error){
-
-        return res.status(400).send({ data: { errorMessage: "Server failed to delete user"}})
+        return res
+            .status(200)
+            .send({ data: { successMessage: 'User deleted' } });
+    } catch (error) {
+        return res
+            .status(400)
+            .send({ data: { errorMessage: 'Server failed to delete user' } });
     }
-
-
-
-
 });
-
 
 export default router;
